+++ packages/wavs/src/engine/runner/generic.rs
@@ -50,6 +50,8 @@ pub trait EngineRunner: Send + Sync {
             self.engine()
                 .execute(component, workflow.fuel_limit, action, &service.config)?;
 
+        // If Ok(Some(x)), send the result down the pipeline to the submit processor
+        // If Ok(None), just end early here, performing no action (but updating local state if needed)
         if let Some(wasi_result) = wasi_result {
             let service_id = trigger_config.service_id.clone();
             let workflow_id = trigger_config.workflow_id.clone();