+++ examples/components/echo-data/src/lib.rs
@@ -5,7 +5,7 @@ use example_helpers::trigger::{decode_trigger_event, encode_trigger_output};
 struct Component;
 
 impl Guest for Component {
-    fn run(trigger_action: TriggerAction) -> std::result::Result<Vec<u8>, String> {
+    fn run(trigger_action: TriggerAction) -> std::result::Result<Option<Vec<u8>>, String> {
         let (trigger_id, req) =
             decode_trigger_event(trigger_action.data).map_err(|e| e.to_string())?;
 
@@ -13,13 +13,16 @@ impl Guest for Component {
             if input_str.contains("envvar:") {
                 let env_var = input_str.split("envvar:").nth(1).unwrap();
                 if let Ok(value) = std::env::var(env_var) {
-                    return Ok(encode_trigger_output(trigger_id, value.as_bytes().to_vec()));
+                    return Ok(Some(encode_trigger_output(
+                        trigger_id,
+                        value.as_bytes().to_vec(),
+                    )));
                 } else {
                     return Err(format!("env var {} not found", env_var));
                 }
             }
         }
-        Ok(encode_trigger_output(trigger_id, req))
+        Ok(Some(encode_trigger_output(trigger_id, req)))
     }
 }