+++ packages/cli/src/command/exec_component.rs
@@ -12,7 +12,7 @@ use crate::{
 };
 
 pub struct ExecComponent {
-    pub output_bytes: Vec<u8>,
+    pub output_bytes: Option<Vec<u8>>,
     pub fuel_used: u64,
 }
 
@@ -20,14 +20,19 @@ impl std::fmt::Display for ExecComponent {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "Fuel used: \n{}", self.fuel_used)?;
 
-        write!(
-            f,
-            "\n\nResult (hex encoded): \n{}",
-            const_hex::encode(&self.output_bytes)
-        )?;
+        match &self.output_bytes {
+            Some(bytes) => {
+                write!(
+                    f,
+                    "\n\nResult (hex encoded): \n{}",
+                    const_hex::encode(bytes)
+                )?;
 
-        if let Ok(s) = std::str::from_utf8(&self.output_bytes) {
-            write!(f, "\n\nResult (utf8): \n{}", s)?;
+                if let Ok(s) = std::str::from_utf8(bytes) {
+                    write!(f, "\n\nResult (utf8): \n{}", s)?;
+                }
+            }
+            None => write!(f, "\n\nResult: None")?,
         }
 
         Ok(())
@@ -147,7 +152,7 @@ mod test {
 
         let result = ExecComponent::run(&Config::default(), args).await.unwrap();
 
-        assert_eq!(result.output_bytes, b"hello world");
+        assert_eq!(result.output_bytes.unwrap(), b"hello world");
         assert!(result.fuel_used > 0);
 
         // Same idea but hex-encoded with prefix
@@ -160,7 +165,7 @@ mod test {
 
         let result = ExecComponent::run(&Config::default(), args).await.unwrap();
 
-        assert_eq!(result.output_bytes, b"hello world");
+        assert_eq!(result.output_bytes.unwrap(), b"hello world");
         assert!(result.fuel_used > 0);
 
         // Do not hex-decode without the prefix
@@ -173,7 +178,7 @@ mod test {
 
         let result = ExecComponent::run(&Config::default(), args).await.unwrap();
 
-        assert_eq!(result.output_bytes, b"68656C6C6F20776F726C64");
+        assert_eq!(result.output_bytes.unwrap(), b"68656C6C6F20776F726C64");
         assert!(result.fuel_used > 0);
 
         // And filepath
@@ -190,7 +195,7 @@ mod test {
 
         let result = ExecComponent::run(&Config::default(), args).await.unwrap();
 
-        assert_eq!(result.output_bytes, b"hello world");
+        assert_eq!(result.output_bytes.unwrap(), b"hello world");
         assert!(result.fuel_used > 0);
     }
 }